prototype/requirements-appinsights.txt
azure-monitor-opentelemetry-exporter==1.0.0b34
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-exporter-otlp-proto-http==1.29.0
protobuf==5.29.5
python-dotenv==1.2.1
# For trace viewer
flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0
